renames topics to sections in filesystem
