+++ packages/cms/src/index.js
@@ -2,5 +2,5 @@ export {default as Builder} from "./Builder.jsx";
 
 // Components
 export {default as Profile} from "./components/Profile.jsx";
-export {default as Topic} from "./components/Topic.jsx";
+export {default as Section} from "./components/Section.jsx";
 export {default as Search} from "./components/Search/Search.jsx";
