+++ packages/cms/src/profile/DimensionBuilder.jsx
@@ -1,7 +1,7 @@
 import React, {Component} from "react";
 import DimensionCard from "../components/cards/DimensionCard";
 import DimensionCreator from "../components/DimensionCreator";
-import Section from "../components/Section";
+import Panel from "../components/Panel";
 import {Dialog} from "@blueprintjs/core";
 
 import "./DimensionBuilder.css";
@@ -37,7 +37,7 @@ export default class DimensionBuilder extends Component {
 
     return (
       <React.Fragment>
-        <Section
+        <Panel
           title="Dimensions"
           entity="dimension"
           addItem={() => this.setState({isOpen: !this.state.isOpen})}
